remove unused composables
